+++ tests/test_pd_controller.cpp
@@ -3,14 +3,20 @@
 #include "Eigen/Core"
 #include "gtest/gtest.h"
 #include "linear_feedback_controller/pd_controller.hpp"
+#include "utils/mock_robot_model_builder_smart.hpp"
 
 // Basic fixture
 class PDControllerTest : public ::testing::Test {
  protected:
   void SetUp() override {
+    mock_rmb_ = std::make_shared<
+        linear_feedback_controller::SmartMockRobotModelBuilder>();
     controller_ = std::make_unique<linear_feedback_controller::PDController>();
+    controller_->initialize(mock_rmb_);
   }
 
+  std::shared_ptr<linear_feedback_controller::SmartMockRobotModelBuilder>
+      mock_rmb_;
   std::unique_ptr<linear_feedback_controller::PDController> controller_;
 };
 
@@ -44,28 +50,35 @@ TEST_F(PDControllerTest, SetGainsWithStdVector) {
 
 // Verify error on gain size
 TEST_F(PDControllerTest, SetGainsWithMismatchedSizesThrows) {
-  Eigen::VectorXd p_gains(2);
-  Eigen::VectorXd d_gains(3);
-  EXPECT_THROW(controller_->set_gains(p_gains, d_gains), std::invalid_argument);
+  // p_gains have wrong size (nv = 2, we pass 3)
+  Eigen::VectorXd p_gains_bad(3);
+  Eigen::VectorXd d_gains_ok(2);
+  EXPECT_THROW(controller_->set_gains(p_gains_bad, d_gains_ok),
+               std::invalid_argument);
 
-  std::vector<double> p_gains_vec = {1.0, 2.0};
-  std::vector<double> d_gains_vec = {1.0};
-  EXPECT_THROW(controller_->set_gains(p_gains_vec, d_gains_vec),
+  std::vector<double> p_gains_vec_bad = {1.0, 2.0, 3.0};
+  std::vector<double> d_gains_vec_ok = {1.0, 2.0};
+  EXPECT_THROW(controller_->set_gains(p_gains_vec_bad, d_gains_vec_ok),
                std::invalid_argument);
 }
 
 // Verify error on references size
 TEST_F(PDControllerTest, SetReferenceWithMismatchedSizesThrows) {
-  Eigen::VectorXd tau_ref(2);
-  Eigen::VectorXd q_ref(3);
-  EXPECT_THROW(controller_->set_reference(tau_ref, q_ref),
+  Eigen::VectorXd tau_ref_bad(3);  // wrong size (nv = 2)
+  Eigen::VectorXd q_ref_ok(2);
+  EXPECT_THROW(controller_->set_reference(tau_ref_bad, q_ref_ok),
+               std::invalid_argument);
+
+  Eigen::VectorXd tau_ref_ok(2);
+  Eigen::VectorXd q_ref_bad(3);  // mauvaise taille (nq = 2)
+  EXPECT_THROW(controller_->set_reference(tau_ref_ok, q_ref_bad),
                std::invalid_argument);
 }
 
 // Verify error on NaN and INF gains values using Eigen
 TEST_F(PDControllerTest, SetGainsWithSpecialValuesThrows) {
-  Eigen::VectorXd p_gains = Eigen::VectorXd::Ones(3);
-  Eigen::VectorXd d_gains = Eigen::VectorXd::Ones(3);
+  Eigen::VectorXd p_gains = Eigen::VectorXd::Ones(2);
+  Eigen::VectorXd d_gains = Eigen::VectorXd::Ones(2);
 
   double inf = std::numeric_limits<double>::infinity();
   double nan = std::numeric_limits<double>::quiet_NaN();
@@ -76,14 +89,14 @@ TEST_F(PDControllerTest, SetGainsWithSpecialValuesThrows) {
 
   // Test with Infinity
   p_gains(1) = 1.0;  // reset
-  d_gains(2) = inf;
+  d_gains(1) = inf;
   EXPECT_THROW(controller_->set_gains(p_gains, d_gains), std::invalid_argument);
 }
 
 // Verify error on NaN and INF gains values using vectors
 TEST_F(PDControllerTest, SetGainsFromStdVectorWithSpecialValuesThrows) {
-  std::vector<double> p_gains_vec = {1.0, 1.0, 1.0};
-  std::vector<double> d_gains_vec = {1.0, 1.0, 1.0};
+  std::vector<double> p_gains_vec = {1.0, 1.0};
+  std::vector<double> d_gains_vec = {1.0, 1.0};
 
   double inf = std::numeric_limits<double>::infinity();
   double nan = std::numeric_limits<double>::quiet_NaN();
@@ -95,15 +108,15 @@ TEST_F(PDControllerTest, SetGainsFromStdVectorWithSpecialValuesThrows) {
 
   // Test with Infinity
   p_gains_vec[1] = 1.0;  // reset
-  d_gains_vec[2] = inf;
+  d_gains_vec[1] = inf;
   EXPECT_THROW(controller_->set_gains(p_gains_vec, d_gains_vec),
                std::invalid_argument);
 }
 
 // Verify error on NaN and INF references values
 TEST_F(PDControllerTest, SetReferenceWithSpecialValuesThrows) {
-  Eigen::VectorXd q_ref = Eigen::VectorXd::Zero(3);
-  Eigen::VectorXd tau_ref = Eigen::VectorXd::Zero(3);
+  Eigen::VectorXd q_ref = Eigen::VectorXd::Zero(2);
+  Eigen::VectorXd tau_ref = Eigen::VectorXd::Zero(2);
 
   double inf = std::numeric_limits<double>::infinity();
   double nan = std::numeric_limits<double>::quiet_NaN();
@@ -118,34 +131,39 @@ TEST_F(PDControllerTest, SetReferenceWithSpecialValuesThrows) {
                std::invalid_argument);
 }
 
-// TODO : empty_vec -> should it be checked ?
-TEST_F(PDControllerTest, SetGainsWithEmptyVectors) {
+// Verify eror on set gains with empty vectors
+TEST_F(PDControllerTest, SetGainsWithEmptyVectorsThrows) {
   Eigen::VectorXd empty_vec;
-
-  EXPECT_NO_THROW(controller_->set_gains(empty_vec, empty_vec));
+  EXPECT_THROW(controller_->set_gains(empty_vec, empty_vec),
+               std::invalid_argument);
 }
 
 // Fixture for computation tests.
 class PDControllerComputationTest : public ::testing::Test {
  protected:
   void SetUp() override {
+    mock_rmb_ = std::make_shared<
+        linear_feedback_controller::SmartMockRobotModelBuilder>();  // nv = nq =
+                                                                    // 2
     controller_ = std::make_unique<linear_feedback_controller::PDController>();
-    dof_ = 3;
-
+    controller_->initialize(mock_rmb_);
+    dof_ = mock_rmb_->get_nv();
     p_gains_.resize(dof_);
     d_gains_.resize(dof_);
     tau_ref_.resize(dof_);
     q_ref_.resize(dof_);
 
-    p_gains_ << 10.0, 20.0, 30.0;
-    d_gains_ << 1.0, 2.0, 3.0;
-    tau_ref_ << 5.0, 6.0, 7.0;
-    q_ref_ << 1.0, 1.0, 1.0;
+    p_gains_ << 10.0, 20.0;
+    d_gains_ << 1.0, 2.0;
+    tau_ref_ << 5.0, 6.0;
+    q_ref_ << 1.0, 1.0;
 
     controller_->set_gains(p_gains_, d_gains_);
     controller_->set_reference(tau_ref_, q_ref_);
   }
 
+  std::shared_ptr<linear_feedback_controller::SmartMockRobotModelBuilder>
+      mock_rmb_;
   std::unique_ptr<linear_feedback_controller::PDController> controller_;
   int dof_;
   Eigen::VectorXd p_gains_, d_gains_, tau_ref_, q_ref_;
@@ -155,19 +173,17 @@ class PDControllerComputationTest : public ::testing::Test {
 TEST_F(PDControllerComputationTest, ComputeControlMultiDimensional) {
   Eigen::VectorXd q(dof_);
   Eigen::VectorXd v(dof_);
-  q << 0.5, 0.8, 0.9;
-  v << 0.1, -0.2, 0.3;
+  q << 0.5, 0.8;
+  v << 0.1, -0.2;
 
   const Eigen::VectorXd& tau = controller_->compute_control(q, v);
 
   // tau = tau_ref + Kp * (q_ref - q) - Kd * v
   // tau[0] = 5.0 + 10.0 * (1.0 - 0.5) - 1.0 * 0.1   = 5.0 + 5.0 - 0.1 = 9.9
   // tau[1] = 6.0 + 20.0 * (1.0 - 0.8) - 2.0 * (-0.2) = 6.0 + 4.0 + 0.4 = 10.4
-  // tau[2] = 7.0 + 30.0 * (1.0 - 0.9) - 3.0 * 0.3   = 7.0 + 3.0 - 0.9 = 9.1
   ASSERT_EQ(tau.size(), dof_);
   EXPECT_NEAR(tau(0), 9.9, 1e-9);
   EXPECT_NEAR(tau(1), 10.4, 1e-9);
-  EXPECT_NEAR(tau(2), 9.1, 1e-9);
 
   // Check that method return only a reference of internal vector.
   const Eigen::VectorXd& tau2 = controller_->compute_control(q, v);