+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Releases are available on the [github repository](https://github.com/loco-3d/lin
 
 ## [Unreleased]
 
+- PDController: use pinocchio::difference in pd controller to handle SO2 joint
+
 ## [3.2.0] - 2026-04-09
 
 - implement continuous joint support